adding condition for file deletion

+++ init.d/atl-init-confluence.sh
@@ -86,7 +86,9 @@ function configureSharedHome {
         atl_ChangeFolderOwnership "${ATL_CONFLUENCE_USER}" "${ATL_CONFLUENCE_USER}" "${CONFLUENCE_SHARED}"
         su "${ATL_CONFLUENCE_USER}" -c "ln -s \"${CONFLUENCE_SHARED}\" \"${ATL_CONFLUENCE_SHARED_HOME}\"" >> "${ATL_LOG}" 2>&1
         if [ "x${ATL_STANDALONE_MODE}" == "xtrue" ] ; then
-            rm ${CONFLUENCE_SHARED}/cluster.properties
+            if [ -f "${CONFLUENCE_SHARED}/cluster.properties" ] ; then
+                rm ${CONFLUENCE_SHARED}/cluster.properties
+            fi
         fi
     else
         atl_log "No mountpoint for shared home exists. Failed to create cluster.properties file."
